useBreakpoint extracted to hooks.ts and made public for the lib
