+++ lib/hooks.ts
@@ -1,0 +1,18 @@
+import { useEffect, useState } from "react";
+
+export function useBreakpoint(query: string) {
+  const [matches, setMatches] = useState(() => window.matchMedia(query).matches);
+
+  useEffect(() => {
+    const media = window.matchMedia(query);
+    const listener = () => setMatches(media.matches);
+    media.addEventListener("change", listener);
+    return () => media.removeEventListener("change", listener);
+  }, [query]);
+
+  return matches;
+}
+
+export function useIsSmallBreakpoint() {
+  return useBreakpoint("(max-width: 640px)");
+}
