+++ lib/index.ts
@@ -1,3 +1,5 @@
 import "./global.css";
 
 export * from "./components";
+export * from "./hooks";
+export * from "./utils";
